Fix XSS vulnerability.
Technically, if freewall is used on a page where users can contribute content, especially HTML, if a user knows the site is running freewall he/she can insert any arbitrary javascript in the following way:

```html
<div data-fixSize="alert('xss!')"></div>
```

This can easily be fixed by using `$.fn.data()`, which is available in both jQuery and Zepto. It converts "true", "false", and "null" to the correct types, and numbers to the Number type, allowing the conditional on L76 to work.